combined BTModel into DataModel
